Reified compilation modes
