+++ lib/gobstones.rb
@@ -8,5 +8,6 @@ end
 
 require_relative './gobstones/batch'
 require_relative './gobstones/batch_parser'
+require_relative './gobstones/compilation_mode'
 require_relative './gobstones/error_builder'
 require_relative './gobstones/example_code_builder'
