+++ lib/gobstones/batch.rb
@@ -1,11 +1,13 @@
 class Gobstones::Batch
   attr_accessor :options, :examples, :content, :extra
+  delegate :compile_content, :compile_extra, to: :@compilation_mode
 
   def initialize(content, examples, extra, options)
     @content = content
     @examples = examples
     @extra = extra
     @options = options
+    @compilation_mode = @options[:game_framework] ? Gobstones::CompilationMode::GameFramework : Gobstones::CompilationMode::Classic
   end
 
   def run_tests!(output)
@@ -16,52 +18,13 @@ class Gobstones::Batch
 
   def to_json
     {
-      code: make_content,
-      extraCode: make_extra,
+      code: compile_content(content),
+      extraCode: compile_extra(extra),
       examples: examples.map { |example| example_json(example) }
     }.to_json
   end
 
   private
-
-  def make_extra
-    if @options[:game_framework]
-      [extra, render_framework_file('extra.gbs')]
-    else
-      extra
-    end
-  end
-
-  def make_content
-    if @options[:game_framework]
-      content_with_framework_program
-    else
-      content
-    end
-  end
-
-  def content_with_framework_program
-    if blockly_content?
-      xml = Nokogiri::XML(content)
-      xml.root.add_child render_framework_file('program.xml')
-      xml.to_xhtml.gsub(/\n\s*/, '')
-    else
-      <<~GBS
-        #{content}
-
-        #{render_framework_file 'program.gbs'}
-      GBS
-        .chop
-    end
-  end
-
-  def blockly_content?
-    Nokogiri::XML(content).errors.empty?
-  end
-
-  def render_framework_file(name)
-    ERB.new(File.read("lib/game_framework/#{name}.erb")).result
-  end
 
   def example_json(example)
     expected_board = example[:postconditions][:final_board]
